LayoutItem to store View's Margin || Padding if Layout set
Prevents the legacy size negotiation repositioning View with Margins
or Padding..
Control-impl Resets position and size if Margin or Padding set.

Solution here is to move Margin and Padding  storage from View to LayoutItem so
native code will not find a margin or padding hence not try to reset positions.

Change-Id: I1d874fa854b0f229059ddfa3476b1426014d075e